+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### WIP
   - Added slidenav-inverted modifier
   - Added direction info to the uk-scroll event
+  - Added $.UIkit.Utils.str2json helper function
   - Updated active paremeter for the switcher component to allow false as value
   - Fixed automute in slideshow component